feat: add functionality to display and add teachers to a list
This commit adds a new file, app.js, which contains the logic to display a list of teachers and add new teachers to the list. The list is displayed in a new section added to the index.html file. The style.css file is also added to style the list of teachers.